Add Gamma single-sampler
Adds a single-sample function for the Gamma family of distributions using K+Theta parameterization.

I validated the prototypes for these in R, and they should work, but I'm not certain that the added complexity of switching between two sampling methods is worth it. Someone should run some profiles, but it probably won't be me just yet.